### 2024-09-04 #7182 @editorialbot
[PRE REVIEW]: ndbc-api: Accelerating oceanography and climate science research with Python
**Submitting author:** @CDJellen (<a href="http://orcid.org/0000-0003-0469-353X">Chris Jellen</a>)
**Repository:** http://github.com/cdjellen/ndbc-api
**Branch with paper.md** (empty if default branch): user/cjellen/joss-paper-submission
**Version:** v2024.08.31.1
