+++ paper.bib
@@ -22,7 +22,8 @@
   number = {1},
   pages = {56-61},
   title = {Data Structures for Statistical Computing in Python},
-  journal = {Proceedings of the 9th Python in Science Conference}
+  journal = {Proceedings of the 9th Python in Science Conference},
+  doi = {10.25080/majora-92bf1922-00a}
 }
 
 @misc{unidata_netcdf4,
