### 2024-09-11 #7182 @kthyng
@CDJellen the references in your paper aren't working correctly. Check that you are following the syntax exactly from https://joss.readthedocs.io/en/latest/paper.html#joss-paper-format
