### 2024-09-12 1d1ccc0
better typing allow lists with only str or int
