### 2024-09-12 #7182 @CDJellen
Thank you @kthyng , I believe I've fixed the error and re-tested through Docker. The rendered version is below.
[paper.pdf](https://github.com/user-attachments/files/16982352/paper.pdf)
