### 2024-09-24 #7182 @fernando-aristizabal
Hi @cheginit,
I'm sorry, but I have a busy schedule for the next few weeks and won't be able to give this submission a proper review. It is an interesting project, and I am open to supporting similar reviews in the future. Perhaps @jarq6c would be interested in this, given his work with [hydrotools](https://github.com/NOAA-OWP/hydrotools)?
