### 2024-10-03 #7182 @jarq6c
> @fernando-aristizabal Thanks for your quick response, letting me know about your availability, and suggesting a potential reviewer! Appreciate it.
>
> @jarq6c Would you like to review this submission to the Journal for Open Source Software? Our reviews are checklist-driven and openly conducted on GitHub over a timeline of 4–6 weeks. Because the process is much more iterative and interactive than a traditional paper review, we would ask you to start within the next 2 weeks. Here are reviewer guidelines for reference: [joss.readthedocs.io/en/latest/reviewer_guidelines.html](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
>
> Thanks for your consideration.

Sorry, I don't think I'm qualified to review this package. It looks like the review guidelines are mainly software focused. I'm not a software developer and my experience is limited to reviewing traditional journal articles in the hydrologic science domain. Best of luck!
