### 2024-10-03 #7182 @cheginit
@jarq6c Thank you for your response. I want to reassure you that being a software developer isn't required for JOSS reviews. As a scientist who uses and develops code in hydrologic research, your expertise is valuable.

We seek reviewers who can assess both scientific merit and practical application. Your experience with traditional journal reviews and domain knowledge makes you well-suited to evaluate JOSS submissions in your field. If you're willing to reconsider, we'd appreciate your contribution. If not, we understand. Please let me know if you have any questions.
