+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "ndbc-api"
-version = "0.24.08.31.1"
+version = "0.24.10.15.1"
 description = "A Python API for the National Data Buoy Center."
 authors = ["cdjellen <cdjellen@gmail.com>"]
 license = "MIT"
