### 2024-10-24 #7406 @editorialbot
[REVIEW]: ndbc-api: Accelerating oceanography and climate science research with Python
**Submitting author:** @CDJellen (<a href="http://orcid.org/0000-0003-0469-353X">Chris Jellen</a>)
**Repository:** http://github.com/cdjellen/ndbc-api
**Branch with paper.md** (empty if default branch): user/cjellen/joss-paper-submission
**Version:** v0.24.12.02.1
