### 2024-11-13 CDJellen/ndbc-api#49 @ks905383
Suggestion: adding station as index to multi-station retrievals
When getting observations from multiple stations:
```
from ndbc_api import NdbcApi

api = NdbcApi()

# get all standard meterological measurements for stations tplm2 and apam2
stdmet_df = api.get_data(
    station_ids=['tplm2', 'apam2'],
    mode='stdmet',
    start_time='2021-01-01',
    end_time='2022-01-01',
)

stdmet_df
```

The index of the returned `df` is only the `timestamp`, resulting in a non-unique index (because timestamps can be duplicated between different stations). Perhaps it may be useful to make `station_id` an index as well.

In addition to better leveraging `pandas`'s index-based functions, this would also make it easier to flip between `pandas` and `xarray` (since `df.to_xarray()` uses those indices), for example.

(it's easy enough to do 'manually' through `stdmet_df.reset_index().set_index(['station_id','timestamp'])`, but it may be a neater way of showing this information)

This is related to JOSS review https://github.com/openjournals/joss-reviews/issues/7406.
