### 2024-11-13 CDJellen/ndbc-api#50 @ks905383
Apologies, like one of the other issues I've opened, this actually seems to have been fixed with the latest version.

I had python==3.13.0 in the environment I was testing this in, so pip and conda installed an older version instead.
