### 2024-11-13 #7406 @ks905383
@CDJellen Could you clarify something on the modes? The docs say that the different modes '[correspond] to the data formats provided by the NDBC data service'.  I'm not quite sure how that maps onto the data. I see that some of the modes carry some of the same variables, but that the same timestamp + station can give different values for those variables (like `WDIR`, `WSPD`, and `GST` in the example above). Would a user more familiar with this dataset know the difference between those two rows, or should there be a flag / column (/ df index) specifying which data format each row came from?
(apologies, I haven't worked with this specific dataset before)

Thanks!
