### 2024-11-15 CDJellen/ndbc-api#51 @CDJellen
Thank you for the suggestion @ks905383 this looks excellent. At a minimum, the switch over to xarray as a wrapper for the netCDF4 datasets is worth implementing. I initially attempted this directly through xarray, but was unable to load the data from URL without the extra steps of reading it to a local temp file. In terms of including the station metadata, the snippet you posted speaks for itself. I'll work on implementing this, possibly with an `include_metadata` flag or similar. I'll also try to make sure the behavior maps well to `xarray.concat`.

The latter might take some time to implement, but the first will be included in the next release.
