+++ tests/api/requests/http/test_swr1.py
@@ -3,8 +3,8 @@ import yaml
 
 from ndbc_api.api.requests.http.swr1 import Swr1Request
 from tests.api.requests.http._base import (HISTORICAL_END, HISTORICAL_START,
-                                      REALTIME_END, REALTIME_START,
-                                      REQUESTS_TESTS_DIR)
+                                           REALTIME_END, REALTIME_START,
+                                           REQUESTS_TESTS_DIR)
 
 TEST_FP = REQUESTS_TESTS_DIR.joinpath('swr1.yml')
 TEST_STN = '41001'
