+++ tests/test_ndbc_api.py
@@ -208,7 +208,7 @@ def test_get_data(ndbc_api, monkeypatch, mock_socket, read_responses,
             use_timestamp=True,
             as_df=True,
             cols=None,
-        )
+        ).reset_index(level='station_id', drop=True)
         pd.testing.assert_frame_equal(
             want[TEST_START:TEST_END].sort_index(axis=1),
             got[TEST_START:TEST_END].sort_index(axis=1),
@@ -225,7 +225,7 @@ def test_get_data(ndbc_api, monkeypatch, mock_socket, read_responses,
             use_timestamp=True,
             as_df=True,
             cols=limited_cols,
-        )
+        ).reset_index(level='station_id', drop=True)
         pd.testing.assert_frame_equal(
             want[TEST_START:TEST_END].sort_index(axis=1),
             got[TEST_START:TEST_END].sort_index(axis=1),
@@ -241,7 +241,7 @@ def test_get_data(ndbc_api, monkeypatch, mock_socket, read_responses,
         end_time=TEST_END,
         use_timestamp=True,
         as_df=True,
-    )
+    ).reset_index(level='station_id', drop=True)
     pd.testing.assert_frame_equal(
         want[TEST_START:TEST_END].sort_index(axis=1),
         got[TEST_START:TEST_END].sort_index(axis=1),
@@ -290,7 +290,7 @@ def test_get_data(ndbc_api, monkeypatch, mock_socket, read_responses,
             as_df=True,
             cols=['FOOBAR'],
         )
-    with pytest.raises(ResponseException):
+    with pytest.raises(RequestException):
         _ = ndbc_api.get_data(
             station_id=globals()[f'TEST_STN_{name.upper()}'],
             mode='foo',
@@ -406,32 +406,3 @@ def test_handle_data(ndbc_api, read_parsed_df):
     assert isinstance(want, dict)
     assert isinstance(got, dict)
     assert len(want.keys()) == len(got.keys())
-
-
-@pytest.mark.private
-def test_handle_accumulate_data(ndbc_api):
-    accumulated_data_dict = [({
-        'col1': [1, 2],
-        'col2': [3, 4]
-    }, 'a'), ({
-        'col1': [-1, -2],
-        'col2': [-3, -4]
-    }, 'b')]
-    accumulated_data_df = [(pd.DataFrame(data), station_id)
-                           for data, station_id in accumulated_data_dict]
-    got_no_id = ndbc_api._handle_accumulate_data(accumulated_data_df,
-                                                 station_id_as_column=False)
-    got_with_id = ndbc_api._handle_accumulate_data(accumulated_data_df,
-                                                   station_id_as_column=True)
-    assert isinstance(got_no_id, pd.DataFrame)
-    assert isinstance(got_with_id, pd.DataFrame)
-    assert got_no_id.shape[0] == got_with_id.shape[0]
-    assert got_no_id.shape[1] != got_with_id.shape[1]  # we have an extra column
-
-    got_no_id = ndbc_api._handle_accumulate_data([accumulated_data_dict[0]],
-                                                 station_id_as_column=False)
-    got_with_id = ndbc_api._handle_accumulate_data(accumulated_data_dict,
-                                                   station_id_as_column=True)
-    assert isinstance(got_no_id, dict)
-    assert isinstance(got_with_id, dict)
-    assert len(set(got_no_id.keys()) & set(got_with_id.keys())) == 0
