+++ tests/api/requests/opendap/test_swden.py
@@ -3,8 +3,8 @@ import yaml
 
 from ndbc_api.api.requests.opendap.swden import SwdenRequest
 from tests.api.requests.opendap._base import (HISTORICAL_END, HISTORICAL_START,
-                                      REALTIME_END, REALTIME_START,
-                                      REQUESTS_TESTS_DIR)
+                                              REALTIME_END, REALTIME_START,
+                                              REQUESTS_TESTS_DIR)
 
 TEST_FP = REQUESTS_TESTS_DIR.joinpath('swden.yml')
 TEST_STN = '41001'
