+++ ndbc_api/api/parsers/opendap/ocean.py
@@ -11,6 +11,7 @@ class OceanParser(BaseParser):
     SPATIAL_DIMS = ['latitude', 'longitude', 'instrument', 'water_depth']
 
     @classmethod
-    def nc_from_responses(cls, responses: List[dict], use_timestamp: bool = False) -> pd.DataFrame:
-        return super(OceanParser,
-                     cls).nc_from_responses(responses)
+    def nc_from_responses(cls,
+                          responses: List[dict],
+                          use_timestamp: bool = False) -> pd.DataFrame:
+        return super(OceanParser, cls).nc_from_responses(responses)
