### 2024-11-17 b1b2a56
migrate off of netcdf4 package
