+++ tests/api/parsers/opendap/test_adcp.py
@@ -1,4 +1,4 @@
-import netCDF4 as nc
+import xarray
 
 import pytest
 
