+++ ndbc_api/utilities/opendap/dataset.py
@@ -2,7 +2,6 @@ from datetime import datetime
 from typing import List, Union
 
 import xarray
-import netCDF4 as nc
 import numpy as np
 
 
