+++ ndbc_api/api/parsers/opendap/stdmet.py
@@ -1,6 +1,6 @@
 from typing import List
 
-import netCDF4 as nc
+import xarray
 
 from ndbc_api.api.parsers.opendap._base import BaseParser
 
@@ -13,5 +13,5 @@ class StdmetParser(BaseParser):
     @classmethod
     def nc_from_responses(cls,
                           responses: List[dict],
-                          use_timestamp: bool = False) -> 'nc.Dataset':
+                          use_timestamp: bool = False) -> xarray.Dataset:
         return super(StdmetParser, cls).nc_from_responses(responses)
