+++ tests/api/parsers/opendap/test_wlevel.py
@@ -1,4 +1,4 @@
-import netCDF4 as nc
+import xarray
 
 import pytest
 import yaml
