+++ requirements.txt
@@ -1,6 +1,6 @@
 requests>=2.10.0
-pandas
+pandas>2.0.0
 beautifulsoup4
 html5lib>=1.1
-netcdf4==1.7.1.post2
 xarray
+scipy
