+++ pyproject.toml
@@ -10,10 +10,10 @@ readme = "README.md"
 python = ">=3.8,<3.13"
 requests = ">=2.10.0"
 pandas = ">=1.5.3"
-numpy  = "^1.24.3"
+numpy  = ">=1.24.3"
+scipy = ">=1.7.3"
 beautifulsoup4 = "~4"
 html5lib = "^1.1"
-netcdf4 = "^1.7.1.post2"
 xarray = ">=2022.6.0"
 
 [tool.poetry.group.dev.dependencies]
