### 2024-11-17 fe93915
updates to test CI
