+++ tests/api/parsers/opendap/test_stdmet.py
@@ -18,7 +18,7 @@ def stdmet_response():
 
 @pytest.fixture
 def parsed_stdmet():
-    ds = nc.Dataset(PARSED_FP, 'r')
+    ds = xarray.open_dataset(PARSED_FP, 'r')
     yield ds
 
 
