+++ tests/api/parsers/opendap/test_pwind.py
@@ -18,7 +18,7 @@ def pwind_response():
 
 @pytest.fixture
 def parsed_pwind():
-    ds = nc.Dataset(PARSED_FP, 'r')
+    ds = xarray.open_dataset(PARSED_FP, 'r')
     yield ds
 
 
