### 2024-11-17 4d46d84
adjust test for new api surface
