+++ tests/api/parsers/opendap/test_cwind.py
@@ -18,7 +18,7 @@ def cwind_response():
 
 @pytest.fixture
 def parsed_cwind():
-    ds = xarray.open_dataset(PARSED_FP, 'r')
+    ds = xarray.open_dataset(PARSED_FP)
     yield ds
 
 
