### 2024-11-18 CDJellen/ndbc-api#51 @CDJellen
Thank you again for these recommendations! The changes are included in #52 and the release to PyPi and Conda is from #53.
