### 2024-11-18 CDJellen/ndbc-api#49 @CDJellen
As for #51 , the suggestions are incorporated in #52 and the release in #53. I am scoping an appropriate change which allows for station metadata in the returned xarray Datasets.  Thanks again for the suggestions!
