### 2024-11-19 #7406 @ks905383
Great! I only have two minor comments left (listed below), otherwise I recommend acceptance. Thanks for the work - it's always great to improve access to the often very janky online datastores of climate datasets....

1. Could you add `conda`/`mamba` install instructions to the README? Since it's on `conda-forge` anyways, might as well advertise it.
2. Would you consider making python 3.13 supported? I don't think too much changed that would mess with dependencies / tests (though would be easy to check by running the test package on 3.13 as well), but empty new environments are installing that python version, and if `ndbc-api` isn't specified in the initial environment, it'll install an older version instead (presumably the last one that didn't explicitly specify python requirements).
