+++ pyproject.toml
@@ -7,7 +7,7 @@ license = "MIT"
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = ">=3.8,<3.14"
+python = ">=3.8,<3.13"
 requests = ">=2.10.0"
 pandas = ">=1.5.3"
 numpy  = ">=1.24.3"
