### 2024-11-19 055c165
updates for python 3.12
