### 2024-11-19 #7406 @CDJellen
Thank you @ks905383 ! I've updated the README with `conda` instructions; great point there as this was a relatively recent change.

With respect to python 3.13 support, I've spent some time investigating this but it seems some dependencies (at least as reported through poetry) lack support for a wide enough version set. With that said, I've updated CI to include python 3.12 support in tests, and updated the package dependencies for 3.12 compatibility.

I will revisit support for 3.13 in a few weeks once the package dependencies have more recent updates.

Once again, thank you for taking the time to review and offer feedback on the package. Your suggestions were excellent, the migration to `xarray` is especially useful given how much cleaner that API is when compared to `netCDF4`. Have an excellent rest of your day!
