### 2024-11-21 #7406 @rwegener2
Hi @CDJellen, this is a great package you've built! Python-based access to NDBC data will help lots of folks use their data more easily.

I have comments on two aspects of this project so far:
### Tests

- When I run the tests I get 21 passed, 99 skipped, and 1 warning. The only two test files that ran were `tests/config/test_config.py` and `tests/test_ndbc_api.py`. Is there a reason most of the tests are skipping, or am I making a mistake in how I'm running them?

### Paper

Well written paper! A few points of feedback:
- Line 10: The phrase "file-based access methods" confused me a bit, but made more sense once I looked at your code. To make this clearer from the perspective of someone only reading the paper I suggest either 1) framing the challenge with the "file-based methods" as the fact that NDBC only provides an API but no Python wrapper and/or that the ASCII text files are not easy to parse, or 2) specifying in the next sentence that this package provides a user-friendly Python API that provides direct access to data as Python data structures.
- Line 26: You're addressing an important problem here but I think it gets glazed over in "the mode of access adds cost and complexity to their workflows". I'd either clarify how the current system adds cost (or remove that part) and also specify how it adds complexity. Personally, I would also note that there isn't a Python interface, only an API (as I understand it), downloading otherwise happens in a GUI, and that it is difficult to parse the text files.
- Line 29: typo "these critical gap"
- Line 33: I struggle with the word "modalities" as it is quite vague. I suggest either 1) after the word modalities include examples or a definition (maybe: "... stations, data modalitites (ex. wind speed, ocean temperature, etc.), ..." or 2) switching to another word/phrase such as 'data variables' or 'measurements'
- Line 34: with the suggestion from the previous reviewer should an xarray `DataSet` also be part of this list or objects?
- Line 34: The pandas team requests that you include the zenodo repository in addition to the McKinney article when citing pandas https://pandas.pydata.org/about/citing.html
- Line 47: typo, missing 'and' in "... their teams, their network ..."
