### 2024-11-22 f0125a7
Add test instructions to contributing
