### 2024-11-22 88aafb0
Fix typo in Station measurements
