### 2024-11-23 0611354
Add new CONTRIBUTING.md guide
