### 2024-11-23 ceade1b
improve the introduction for the API in the README
