### 2024-11-23 b99e214
add mode descriptions
