### 2024-11-23 CDJellen/ndbc-api#57 @rwegener2
Example Notebook Clarifications/Suggestions
The example notebook is really helpful for learning about the tool! A few suggestions to make it even clearer:

1. When using the `api.radial_search()` method I had to get an error message to figure out what the options are for the `units` parameter. A sentence stating these three options and what they correspond to would be great.
2. In the "Obtain measurements over a given time range" section there is the following sentence: "The returned pd.DataFrame will match the format explored earlier, with one key difference. The station_id is now included as a column in addition to the other features.". It looks like ‘station’ is a mulit-index, not a column. If that's true this sentence can be updated.
3. I like the paragraph in "Check the API's supported data formats" that gives some context about modes, since modes seems to be a level of organization unique to NDBC. Some pieces of info that I would add to this paragraph (assuming I've understood correctly) are 1) the data variables that are available at a station depends on which modes the station is operating in 2) a station can be operating in multiple simultaneous modes 3) you can't search by data variable, you can only search by mode. If there is a way for people to figure out what data variables correspond to each node that may also be helpful. Maybe consider a table displaying the information from the top of section 8 in the [NDBC data guide](https://www.ndbc.noaa.gov/docs/ndbc_web_data_guide.pdf) which gives a plain language description of the mode codes (ex. `stdmet` -> Standard Meteorological data)
4. Your example notebook has a really nice flow, but it feels like it gets interrupted in the middle with the `api.available_realtime()` and `api.available_historic()` methods. As far as I can tell those methods weren't important for me getting data into pandas, they were sort of just bonus methods to generate access links to the raw data in case I wanted to download the files. I would suggest moving these methods to the end of the notebook to keep the workflow that I would guess is most likely to be used together(find a station -> access station metadata -> access station data).
5. In the notebook as it is I don't actually see how a user would figure out which modes are available at a station. The `api.available_historical()` method gives the plain language description of the mode, but not the code. I realized later that the `api.get_modes()` method accepts a station key. Do you think showing users how to use `.get_modes()` would be a more streamlined way of showing users how to get the station code?

Feel free to disagree with me here, these are just some thoughts after reading through the example notebook with fresh eyes. Please don't hesitate to ask follow ups if anything is unclear!

https://github.com/openjournals/joss-reviews/issues/7406
