### 2024-11-23 #7406 @rwegener2
Thanks for the clarification about tests @CDJellen.

I have left all my comments from the review in the PR/issue above. The software works really nicely, the comments are mostly about bumps I had as I was learning to use the software from the docs and example notebook. Feel free to ping me here or directly in the issues if anything is unclear!
