+++ paper.bib
@@ -26,9 +26,25 @@
   doi = {10.25080/majora-92bf1922-00a}
 }
 
-@misc{unidata_netcdf4,
-  author = {{Unidata}},
-  title = {NetCDF4 Python Library},
-  howpublished = {\url{https://unidata.github.io/netcdf4-python/}},
-  note = {Accessed: 2-September-2024}
+@software{reback2020pandas,
+    author       = {The pandas development team},
+    title        = {pandas-dev/pandas: Pandas},
+    month        = feb,
+    year         = 2020,
+    publisher    = {Zenodo},
+    version      = {latest},
+    doi          = {10.5281/zenodo.3509134},
+    url          = {https://doi.org/10.5281/zenodo.3509134}
+}
+
+@article{hoyer2017xarray,
+  title     = {xarray: {N-D} labeled arrays and datasets in {Python}},
+  author    = {Hoyer, S. and J. Hamman},
+  journal   = {Journal of Open Research Software},
+  volume    = {5},
+  number    = {1},
+  year      = {2017},
+  publisher = {Ubiquity Press},
+  doi       = {10.5334/jors.148},
+  url       = {https://doi.org/10.5334/jors.148}
 }
