### 2024-11-24 CDJellen/ndbc-api#57 @CDJellen
Thank you again for your review and feedback @rwegener2; I've updated the overview notebook and some of the method docstrings based on these suggestions.

The inclusion of the `available_historical` and `available_realtime` methods in the overview notebook is intended to answer question 5 above. Without speculatively making an API call, the easiest way to determine what `mode`s are available at a given station is to use one of these two methods for the station of interest based on the time period required. The responses do include the `mode` indicators, but only through URLs.

I believe that adding either a `get_modes(station_id=...)` or refactoring the response of these two methods will help make this more intuitive and fit-for-purpose.

One of these two will be included as a feature in the next release, along with updates to the flags and keywords across some of the data retrieval methods and possibly the inclusion of station metadata in `opendap` data response.

All other changes were made in #58 . If you would prefer to see these changes expediated, please feel free to re-open this issue. Thank you again and have an excellent rest of your day.
