### 2024-11-26 CDJellen/ndbc-api#57 @rwegener2
All sounds good. I see you added a sentence explaining how that the indicator is in the url, (_"In the response above, the modes which are available for station "tplm2" are reported in plain text. Their indicators (stdmet, cwind) are also specified as the suffix of the URLS for the data files."_) , which is super helpful. I think one of those two more in-depth options you listed will really help your users in the next release!
