### 2024-11-26 #7406 @rwegener2
Updates look good @CDJellen! The paper is really nice, the contributing.md is a solid addition, and I think the clarifications in the example notebook will really help your users.

This package looks good to me. @cheginit I recommend acceptance.
