### 2024-12-02 #7406 @cheginit
@CDJellen Thanks for working with the reviewers and addressing their comments. Before, I hand it over to the EiC for the final publication, please address the following:

Paper:

- [ ] Please change "meterological" to "meteorological". Note that there are two instances.
- [ ] There are several instances of the word "Python" that are not capitalized, e.g., Line 9. There is also one in the references.
- [ ] L10-14: This is a long sentence, please break it down to two.
- [ ] In the references section, the Pandas item should be `pandas/pandas-dev - Pandas` and its author should be `The pandas development team` (you need to add additional curly braces to this item in the bib file.)

Regarding the branch, I noticed that there are two JOSS related branches. Once you addressed my comments, please make sure to release a new version with all the commits that you want to be included in your JOSS publication. Then provide me its version number and an archive DOI (e.g., on Zenodo).
