+++ paper.md
@@ -18,13 +18,13 @@ bibliography: paper.bib
 
 # Summary
 
-The National Data Buoy Center (NDBC) and its partners are an essential source of marine meteorological and oceanographic data [@NDBC_Data_Guide]. The `ndbc-api` Python package is an open-source tool designed to streamline the acquisition, synthesis, and analysis of this data. It provides a python API for accessing real-time and historical observations from a network of buoys, coastal stations, and deployments. This package simplifies the process of retrieving, parsing, and organizing NDBC data, particularly when dealing with multiple stations or extended time ranges, which can be cumbersome using the traditional HTTP access methods serving station-specific text files maintained by the NDBC. The `ndbc-api` provides a python API which retrieves, parses, and merges this data, returning python-native data structures. This package empowers researchers and practitioners in oceanography, meteorology, and related fields to efficiently integrate NDBC data into their workflows, accelerating research in climate science and oceanography.
+The National Data Buoy Center (NDBC) and its partners are an essential source of marine meteorological and oceanographic data [@NDBC_Data_Guide]. The `ndbc-api` Python package is an open-source tool designed to streamline the acquisition, synthesis, and analysis of this data. It provides a Python API for accessing real-time and historical observations from a network of buoys, coastal stations, and deployments. This package simplifies the process of retrieving, parsing, and organizing NDBC data. Traditional HTTP access methods to station-specific text files poses a challenge when dealing with multiple stations or extended time ranges. The `ndbc-api` provides a Python API which retrieves, parses, and merges this data, returning Python-native data structures. This package empowers researchers and practitioners in oceanography, meteorology, and related fields to efficiently integrate NDBC data into their workflows, accelerating research in climate science and oceanography.
 
 # Statement of need
 
-The National Oceanic and Atmospheric Association's National Data Buoy Center maintains marine monitoring and observation stations around the world [@NDBC_Active_Stations]. These stations report atmospheric, oceanographic, and other meterological data at regular intervals to the NDBC [@NDBC_Data_Guide;@NDBC_Active_Stations]. Measurements are made available over HTTP through the NDBC's data service. Measurements are typically distributed as quality-controlled `utf-8` encoded, station-by-station, fixed-period text files [@NDBC_Data_Guide]. While the data collected and maintained by the NDBC is critical to oceanography and climate science researchers, the mode of access adds complexity to their workflows. These challenges are particularly pronounced when working with long-duration data, data from multiple stations, or data with a high proportion of missing measurements.
+The National Oceanic and Atmospheric Association's National Data Buoy Center maintains marine monitoring and observation stations around the world [@NDBC_Active_Stations]. These stations report atmospheric, oceanographic, and other meteorological data at regular intervals to the NDBC [@NDBC_Data_Guide;@NDBC_Active_Stations]. Measurements are made available over HTTP through the NDBC's data service. Measurements are typically distributed as quality-controlled `utf-8` encoded, station-by-station, fixed-period text files [@NDBC_Data_Guide]. While the data collected and maintained by the NDBC is critical to oceanography and climate science researchers, the mode of access adds complexity to their workflows. These challenges are particularly pronounced when working with long-duration data, data from multiple stations, or data with a high proportion of missing measurements.
 
[…]
