+++ paper.bib
@@ -28,7 +28,7 @@
 
 @software{reback2020pandas,
     author       = {The pandas development team},
-    title        = {pandas-dev/pandas: Pandas},
+    title        = {{pandas-dev/pandas: Pandas}},
     month        = feb,
     year         = 2020,
     publisher    = {Zenodo},
