### 2024-12-02 #7406 @CDJellen
Thank you again! I've created a new release for version `v0.24.12.02.1` in GitHub, which is released in PyPi and conda forge.

The DOI in Zenodo for the package is ` 10.5281/zenodo.14261354` [10.5281/zenodo.14261354](https://doi.org/10.5281/zenodo.14261354). For this specific release, we have the DOI `10.5281/zenodo.14261355` [10.5281/zenodo.14261355](https://doi.org/10.5281/zenodo.14261355)
