### 2024-12-02 #7406 @cheginit
@CDJellen The issues with the Pandas authors in the references and not capitalized Python in the second item are still remaining. You need to put the author for the Pandas item in curly braces `{{The pandas development team}}` and the same for the second item ` ... computing in {Python} ...`
