+++ paper.bib
@@ -21,13 +21,13 @@
   volume = {445},
   number = {1},
   pages = {56-61},
-  title = {Data Structures for Statistical Computing in Python},
+  title = {{Data Structures for Statistical Computing in Python}},
   journal = {Proceedings of the 9th Python in Science Conference},
   doi = {10.25080/majora-92bf1922-00a}
 }
 
 @software{reback2020pandas,
-    author       = {The pandas development team},
+    author       = {{The pandas development team}},
     title        = {{pandas-dev/pandas: Pandas}},
     month        = feb,
     year         = 2020,
