### 2024-12-02 #7406 @CDJellen
> @CDJellen The issues with the Pandas authors in the references and not capitalized Python in the second item are still remaining. You need to put the author for the Pandas item in curly braces `{{The pandas development team}}` and the same for the second item ` ... computing in {Python} ...`

@cheginit I apologize for the delay in resolving these; I've corrected both by adding an extra layer of curly braces as suggested. These seem to render correctly from local testing. Thank you again and please let me know if there is anything else I can do to assist.
