### 2024-12-02 #7406 @CDJellen
> @CDJellen I just noticed that the title of the [archive](https://zenodo.org/records/14261355) is not the same as this submission, can you please change it?

I've updated the title of the Zenodo submission to reflect the title of the JOSS submission `ndbc-api: Accelerating oceanography and climate science research with Python`
