### 2024-12-02 #7406 @cheginit
@CDJellen Awesome! Thank you for submitting to JOSS and joining our community! I will now hand it over to EiC for final checks and publication.
