### 2024-12-02 #7406 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
